change due to the removing the Juttle argument from the base class
moved things around so we have a better layout with read and write code
split up into separate javascript modules
