+++ lib/index.js
@@ -1,0 +1,9 @@
+function GraphiteBackend(config) {
+    return {
+        name: 'graphite',
+        read: require('./read')(config),
+        write: require('./write')(config)
+    };
+}
+
+module.exports = GraphiteBackend;
