+++ lib/read.js
@@ -4,13 +4,14 @@ var JuttleMoment = require('juttle/lib/moment').JuttleMoment;
 var net = require('net');
 var Promise = require('bluebird');
 var url = require('url');
+var Juttle = require('juttle/lib/runtime').Juttle;
 
-/* globals global */ 
+/* globals global */
 global.Promise = Promise;
 var fetch = require('isomorphic-fetch');
 
-function GraphiteBackend(config, Juttle) {
-    var Read = Juttle.proc.base.extend({
+var Read = function(config) {
+    return Juttle.proc.base.extend({
         sourceType: 'batch',
         procName: 'read-graphite',
 
@@ -26,13 +27,13 @@ function GraphiteBackend(config, Juttle) {
 
             this.from = options.from;
             this.to = options.to;
-        
-            if (!this.from) { 
+
+            if (!this.from) {
                 throw this.compile_error('RT-REQUIRED-OPTION-ERROR', {
                     proc: 'read graphite',
                     option: '-from'
                 });
-            } 
+            }
 
             var from = Math.round(new JuttleMoment(this.from).milliseconds()/1000);
             var to;
@@ -57,19 +58,19 @@ function GraphiteBackend(config, Juttle) {
                     (expression.operator === '==' || expression.operator === '=~')) {
                     var left = expression.left;
                     var right = expression.right;
-                    if (left && right) { 
+                    if (left && right) {
                         if (left.value === 'name' && right.type === 'StringLiteral') {
                             name = right.value;
                         }
                     }
-                } 
+                }
 
-                if (name) { 
+                if (name) {
                     /* jshint evil: true */
-                    this.filter = eval(source); 
+                    this.filter = eval(source);
                 }
             }
-            
+
             // when no name was set then the filter expression is either missing
             // or wrong
             if (!name) {
@@ -95,14 +96,14 @@ function GraphiteBackend(config, Juttle) {
 
             return fetch(this.url, { headers: self.headers })
             .then(function(res) {
-                if (res.status > 200) { 
+                if (res.status > 200) {
                     throw self.runtime_error('RT-INTERNAL-ERROR', {
                         error: res.status + ': ' + res.statusText
                     });
                 }
                 return res.json();
             })
-            .then(function(payload) { 
+            .then(function(payload) {
                 _.each(payload, function(data) {
                     var points = [];
                     var name = data.target;
@@ -139,70 +140,6 @@ function GraphiteBackend(config, Juttle) {
             });
         }
     });
+};
 
-    var Write = Juttle.proc.sink.extend({
-        procName: 'writex-graphite',
-
-        initialize: function(options, params) {
-            // XXX: missing checks for all required config values
-            this.name = 'writex-graphite';
-            var self = this;
-
-            this.socket = net.Socket();
-            this.socket.connect(config.carbon.port, config.carbon.host);
-
-            this.socket.on('end', function() { 
-                // only eof when we've closed the socket
-                self.on_eof(self);
-            });
-
-            this.isDone = new Promise(function(resolve, reject) {
-                self.on_eof = resolve;
-            });
-        },
-
-        process: function(points) {
-            var self = this;
-            var payload = [];
-
-            _.each(points, function(point) { 
-                var timestamp = Math.round(new Date(point.time).getTime()/1000);
-                
-                if (point.name) { 
-                    if (typeof point.value === 'number' && point.value === point.value) {
-                        payload.push(point.name + ' ' + point.value + ' ' + timestamp + '\n');
-                    } else { 
-                        this.trigger('warning', this.runtime_error('RT-FIELD-NOT-FOUND', {
-                            field: 'value'
-                        }));
-                    }
-                } else {
-                    this.trigger('warning', this.runtime_error('RT-FIELD-NOT-FOUND', {
-                        field: 'name'
-                    }));
-                }
-            });
-            
-            self.socket.write(payload.join(''), 'utf-8', function(err) { 
-                if (err) { 
-                    self.trigger('error', err);
-                    this.done();
-                }
-            });
-        },
-
-        eof: function() {
-            this.socket.end();
-            this.done();
-        },
-
-    });
-
-    return {
-        name: 'graphite',
-        read: Read,
-        write: Write
-    };
-}
-
-module.exports = GraphiteBackend;
+module.exports = Read;
